requirements.txt
addict==2.2.1
audiolazy==0.6
certifi==2020.4.5.1
chardet==3.0.4
facenet-pytorch==2.2.9
ffpyplayer==4.3.1
idna==2.9
mmcv==0.4.3
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
PyAudio==0.2.11
pydub==0.23.1
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.8
